+++ Amethyst/AST/ParserErrorHandler.cs
@@ -8,11 +8,11 @@ namespace Amethyst.AST
 	public class ParserErrorHandler(string path, string file, Visitor visitor) : BaseErrorListener, IAntlrErrorListener<int>, IFileHandler
 	{
 		public readonly string Path = path;
+		public readonly string File = file;
 		public readonly Visitor Visitor = visitor;
 		public bool Errored { get; private set; } = false;
 
-		public Dictionary<string, string> Files => new([new KeyValuePair<string, string>(Path, file)]);
-
+		public string GetFile(string path) => path == Path ? File : throw new FileNotFoundException(path);
 		public string MapToPath(string mappedPath) => mappedPath;
 		public string PathToMap(string path) => path;
 
@@ -27,7 +27,7 @@ namespace Amethyst.AST
 			{
 				var loc = Visitor.LocOffset(new Location(Path, line, charPositionInLine + 1));
 				var term = new CompilerMessager(Color.Red);
-				term.Header($"Error at {loc}");
+				term.Header($"Error at [underline]{loc}[/]");
 				term.AddCode(this, new(loc, new(loc.File, loc.Line, loc.Column + length - 1)));
 				term.Final().AddContent($"Syntax error: [turquoise2]{msg.EscapeMarkup()}[/]");