+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 * Removed `.mcfunction` dumping on debug builds.
 * Added `--dump-cmds` to the CLI.
 * Added `--version` to the CLI.
+* Improved error messages.
+* Added source mapping.
+* Added `/reload` error parsing and display to `amethyst run`.
+* Fixed the source location range for inline commands.